+++ src/index.ts
@@ -1,13 +1,4 @@
-import { getCypressCLIBinPath } from './bin-path';
-import { runInContext } from './context';
-export { run, spawn } from './cypress-wrapper';
+import 'source-map-support/register';
 
-// TODO: extract this to a separate package
-export async function getBinPath(
-  cypressPackagePath: string = require.resolve('cypress')
-) {
-  return await runInContext(
-    getCypressCLIBinPath,
-    new Map().set('cypressPackagePath', cypressPackagePath)
-  );
-}
+export { getCypressCLIBinPath } from './bin-path';
+export { run, spawn } from './cypress-wrapper';